pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "monix"
version = "1.0.0"
description = "Web Security Monitoring API"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Monix Team"}
]
keywords = ["security", "monitoring", "web", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]
requires-python = ">=3.8"
dependencies = [
    "psutil>=5.9.0",
    "requests>=2.28.0",
    "dnspython>=2.4.0",
    "beautifulsoup4>=4.12.0",
    "django>=4.2",
    "django-axes>=8.3.1",
    "django-cors-headers>=4.0.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "gunicorn",
    "cryptography>=42.0.0",
    "social-auth-app-django>=4.0.0",
    "PyJWT>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-django>=4.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/dinexh/monix"
Documentation = "https://github.com/dinexh/monix#readme"
Repository = "https://github.com/dinexh/monix"
Issues = "https://github.com/dinexh/monix/issues"

[tool.setuptools.packages.find]
where = ["core"]
include = ["scan_engine*", "reports*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
DJANGO_SETTINGS_MODULE = "config.settings"
pythonpath = ["core"]
addopts = "--cov=scan_engine --cov=reports --cov-report=term-missing"
